requirements.txt
streamlit
boto3
streamlit_option_menu
PyJWT